initial commit
The code you shared seems to be a Python script for copying files from a source directory to a destination directory, and logging information about the copy operations in a SQL database. It uses the glob, os, shutil, sys, datetime, pandas, apscheduler, pyodbc, and json libraries, and also includes the use of threads for copying multiple files concurrently.

The script reads in configuration parameters and file paths from an Excel file and a JSON file, respectively. It then creates a connection to a SQL database, and sets up a job scheduling mechanism using the apscheduler library. The main function start_threading_copy() creates a thread for each file to be copied, and calls the function get_file() for each thread. The function get_file() first checks if there is a new file in the source directory, by comparing the modification date and file size with the most recently copied file in the destination directory. If there is a new file, it copies it to the destination directory and logs information about the operation in the SQL database.